requirements.txt

aiogram==3.19.0
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.0
python-telegram-bot==20.8
